tools: reserve extra guest memory for ACPI from device model
The device model may need extra guest memory to store the generated
data when building guest ACPI. For example, QEMU puts the unpatched
ACPI and BIOSLinkerLoader ROMs in guest memory, which can be patched
and loaded by the guest firmware later. Though the default value can
be implied from the type of device model and the domain config, we
still add an xl domain config 'dm_acpi_size' to allow users to specify
the required size when the default value is not enough.

Signed-off-by: Haozhong Zhang <haozhong.zhang@intel.com>