xen: add DMOP to access the host NVDIMM
The host NVDIMM pages are not mapped to the domain where QEMU is, but
QEMU may need to read/write the host NVDIMM for vNVDIMM, e.g., storing
vNVDIMM labels to the backend host NVDIMM pages. We introduce a DMOP
command XEN_DMOP_pmem_rw to allow QEMU to access the host NVDIMM pages
that are owned by the domain which QEMU serves.